splitup lagerbank.fxml in small FXMLs
- separate views for: account selection, transaction overview, making a transaction
- trying out a mixture of application events and injecting properties (e.g. selectedAccountProperty) to share information

issue #2
