+++ ljw-fx-client/src/main/java/de/ljw/aachen/gui/LagerbankAppConfig.java
@@ -8,8 +8,7 @@ import de.ljw.aachen.account.management.port.in.ReadAccountUseCase;
 import de.ljw.aachen.account.management.port.out.AccountStorePort;
 import de.ljw.aachen.account.management.service.AccountService;
 import de.ljw.aachen.common.EventPort;
-import de.ljw.aachen.gui.controller.CreateUserController;
-import de.ljw.aachen.gui.controller.LagerbankController;
+import de.ljw.aachen.gui.controller.*;
 import de.ljw.aachen.lagerbank.adapter.out.TransactionStoreMem;
 import de.ljw.aachen.lagerbank.domain.Money;
 import de.ljw.aachen.lagerbank.domain.Transaction;
@@ -17,41 +16,68 @@ import de.ljw.aachen.lagerbank.port.in.*;
 import de.ljw.aachen.lagerbank.port.out.TransactionStorePort;
 import de.ljw.aachen.lagerbank.service.BalanceService;
 import de.ljw.aachen.lagerbank.service.TransactionService;
+import javafx.beans.property.ListProperty;
+import javafx.beans.property.ObjectProperty;
+import javafx.beans.property.SimpleListProperty;
+import javafx.beans.property.SimpleObjectProperty;
+import javafx.collections.FXCollections;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 
 @Configuration
 public class LagerbankAppConfig {
 
     @Bean
-    LagerbankController lagerbankController(ApplicationContext applicationContext,
-                                            ReadAccountUseCase readAccountUseCase,
-                                            ListAccountsUseCase listAccountsUseCase,
-                                            DepositMoneyUseCase depositMoneyUseCase,
-                                            WithdrawMoneyUseCase withdrawMoneyUseCase,
-                                            TransferMoneyUseCase transferMoneyUseCase,
-                                            ListTransactionsUseCase listTransactionsUseCase,
-                                            GetBalanceUseCase getBalanceUseCase) {
-        return new LagerbankController(
-                applicationContext,
-                readAccountUseCase,
-                listAccountsUseCase,
-                depositMoneyUseCase,
-                withdrawMoneyUseCase,
-                transferMoneyUseCase,
-                listTransactionsUseCase,
-                getBalanceUseCase);
+    LagerbankController lagerbankController(ListAccountsUseCase listAccountsUseCase,
+                                            ListProperty<Account> accountListProperty) {
+        return new LagerbankController(listAccountsUseCase, accountListProperty);
     }
 
     @Bean
     CreateUserController createUserController(CreateAccountUseCase createAccountUseCase) {
         return new CreateUserController(createAccountUseCase);
+    }
+
+    @Bean
+    AccountSelectionController accountSelectionController(ApplicationContext applicationContext,
+                                                          ObjectProperty<Account> selectedAccountProperty,
+                                                          ListProperty<Account> accountListProperty) {
+        return new AccountSelectionController(applicationContext, selectedAccountProperty, accountListProperty);
+    }
+
+    @Bean
+    MakeTransactionController makeTransactionController(DepositMoneyUseCase depositMoneyUseCase,
+                                                        WithdrawMoneyUseCase withdrawMoneyUseCase,
+                                                        TransferMoneyUseCase transferMoneyUseCase,
+                                                        ObjectProperty<Account> selectedAccountProperty,
+                                                        ListProperty<Account> accountListProperty) {
+        return new MakeTransactionController(depositMoneyUseCase, withdrawMoneyUseCase, transferMoneyUseCase,
+                selectedAccountProperty, accountListProperty);
+    }
+
+    @Bean
+    TransactionsOverviewController transactionsOverviewController(ListTransactionsUseCase listTransactionsUseCase,
+                                                                  GetBalanceUseCase getBalanceUseCase,
+                                                                  ReadAccountUseCase readAccountUseCase,
+                                                                  ObjectProperty<Account> selectedAccountProperty,
+                                                                  ListProperty<Account> accountListProperty) {
+        return new TransactionsOverviewController(listTransactionsUseCase, getBalanceUseCase,
+                selectedAccountProperty, accountListProperty);
+    }
+
+    @Bean
+    ObjectProperty<Account> selectedAccountProperty() {
+        return new SimpleObjectProperty<>();
+    }
+
+    @Bean
+    ListProperty<Account> accountListProperty() {
+        return new SimpleListProperty<>(FXCollections.observableArrayList());
     }
 
     /* Account management ******************************************************************************************* */
@@ -63,8 +89,9 @@ public class LagerbankAppConfig {
 
     @Bean
     AccountStorePort accountStorePort() {
+        // TODO remove dummy data
         var accounts = List.of(Account.createFor("Peter", "Peterson"), Account.createFor("Julia", "Juliette"));
-        return new AccountStoreMem(accounts);
+        return new AccountStoreMem(accounts); // TODO write to file instead of mem
     }
 
     /* Lagerbank **************************************************************************************************** */
@@ -81,12 +108,12 @@ public class LagerbankAppConfig {
 
     @Bean
     TransactionStorePort transactionStorePort(AccountStorePort accountStorePort) {
-
-        List<Account> accounts = new ArrayList<Account>(accountStorePort.readAll());
+        // TODO remove dummy data
+        var accounts = new ArrayList<>(accountStorePort.readAll());
         var first = accounts.get(0);
         var second = accounts.get(1);
 
-        return new TransactionStoreMem(List.of(
+        return new TransactionStoreMem(List.of( // TODO write to file instead of mem
                 Transaction.forDeposit(first.getId(), Money.of(10)),
                 Transaction.forWithdrawal(first.getId(), Money.of(1)),
                 Transaction.forDeposit(first.getId(), Money.of(1)),
