+++ ljw-fx-client/src/main/resources/fxml/account_selection.fxml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.ListView?>
+<?import javafx.scene.layout.VBox?>
+
+<VBox spacing="5.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="de.ljw.aachen.gui.controller.AccountSelectionController">
+   <children>
+      <ListView fx:id="lvAccounts" minWidth="-Infinity" prefHeight="835.0" prefWidth="247.0" VBox.vgrow="ALWAYS" />
+      <Button fx:id="btnNewUser" maxWidth="1.7976931348623157E308" minHeight="-Infinity" minWidth="-Infinity" mnemonicParsing="false" onAction="#onCreateAccount" text="New account" VBox.vgrow="NEVER" />
+   </children>
+</VBox>
