+++ ljw-fx-client/src/main/resources/fxml/lagerbank.fxml
@@ -1,74 +1,17 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
 <?import javafx.geometry.Insets?>
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.ComboBox?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.ListView?>
-<?import javafx.scene.control.RadioButton?>
 <?import javafx.scene.control.Separator?>
-<?import javafx.scene.control.TableColumn?>
-<?import javafx.scene.control.TableView?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.control.ToggleGroup?>
 <?import javafx.scene.layout.BorderPane?>
-<?import javafx.scene.layout.HBox?>
 <?import javafx.scene.layout.VBox?>
-<?import javafx.scene.text.Font?>
 
 <BorderPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="850.0" prefWidth="1250.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="de.ljw.aachen.gui.controller.LagerbankController">
     <center>
         <VBox spacing="20.0" BorderPane.alignment="TOP_CENTER">
             <children>
-                <TableView fx:id="tvTransactions" prefHeight="350.0" prefWidth="1078.0" tableMenuButtonVisible="true" VBox.vgrow="ALWAYS">
-                    <columns>
-                        <TableColumn fx:id="tcDate" prefWidth="75.0" text="Date" />
-                        <TableColumn fx:id="tcDescription" prefWidth="75.0" text="Description" />
-                        <TableColumn fx:id="tcAmount" prefWidth="75.0" text="Amount" />
-                    </columns>
-                    <columnResizePolicy>
-                        <TableView fx:constant="CONSTRAINED_RESIZE_POLICY" />
-                    </columnResizePolicy>
-                </TableView>
-            <HBox alignment="CENTER_LEFT" spacing="10.0">
-               <children>
-                  <Label text="Total Balance">
-                     <font>
-                        <Font name="System Bold" size="16.0" />
-                     </font>
-                  </Label>
-                  <Label fx:id="lblTotalBalance" text="-">
-                     <font>
-                        <Font name="System Bold" size="16.0" />
-                     </font>
-                  </Label>
-               </children>
-            </HBox>
+                <fx:include source="transactions_overview.fxml" VBox.vgrow="ALWAYS" />
                 <Separator prefWidth="200.0" />
-                <HBox alignment="CENTER_LEFT" spacing="10.0">
-                    <children>
-                        <RadioButton fx:id="rbDeposit" mnemonicParsing="false" text="Deposit">
-                            <toggleGroup>
-                                <ToggleGroup fx:id="tgTransaction" />
-                            </toggleGroup>
-                        </RadioButton>
-                        <RadioButton fx:id="rbWithdraw" mnemonicParsing="false" text="Withdraw" toggleGroup="$tgTransaction" />
-                        <RadioButton fx:id="rbTransfer" mnemonicParsing="false" text="Transfer" toggleGroup="$tgTransaction" />
-                    </children>
-                </HBox>
-                <TextField id="tfAmount" fx:id="tfAmount" promptText="Enter amount" />
-                <ComboBox fx:id="cbReceivers" disable="true" maxWidth="1.7976931348623157E308" promptText="Select recipient">
-                    <padding>
-                        <Insets top="5.0" />
-                    </padding>
-                </ComboBox>
-                <Separator prefWidth="200.0" />
-                <HBox alignment="CENTER" spacing="10.0">
-                    <children>
-                        <Button fx:id="btnApply" defaultButton="true" mnemonicParsing="false" onAction="#onApply" text="Apply" />
-                        <Button fx:id="btnReset" cancelButton="true" mnemonicParsing="false" onAction="#onReset" text="Reset" />
-                    </children>
-                </HBox>
+                <fx:include source="make_transaction.fxml" />
             </children>
             <BorderPane.margin>
                 <Insets left="10.0" />
@@ -76,14 +19,9 @@
         </VBox>
     </center>
     <left>
-        <VBox prefHeight="200.0" prefWidth="100.0" spacing="5.0" BorderPane.alignment="CENTER">
-            <children>
-                <ListView fx:id="lvAccounts" minWidth="-Infinity" prefHeight="835.0" prefWidth="247.0" />
-                  <Button fx:id="btnNewUser" maxWidth="1.7976931348623157E308" minHeight="-Infinity" minWidth="-Infinity" mnemonicParsing="false" onAction="#onCreateAccount" text="New account" VBox.vgrow="NEVER" />
-            </children>
-        </VBox>
+        <fx:include prefHeight="0.0" source="account_selection.fxml" />
     </left>
-   <padding>
-      <Insets bottom="10.0" left="10.0" right="10.0" top="10.0" />
-   </padding>
+    <padding>
+        <Insets bottom="10.0" left="10.0" right="10.0" top="10.0" />
+    </padding>
 </BorderPane>
