+++ ljw-fx-client/src/main/resources/fxml/transactions_overview.fxml
@@ -1,0 +1,33 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.BorderPane?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.VBox?>
+<?import javafx.scene.text.Font?>
+<VBox spacing="20.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="de.ljw.aachen.gui.controller.TransactionsOverviewController">
+    <TableView fx:id="tvTransactions" tableMenuButtonVisible="true"
+               VBox.vgrow="ALWAYS">
+        <columns>
+            <TableColumn fx:id="tcDate" prefWidth="75.0" text="Date"/>
+            <TableColumn fx:id="tcDescription" prefWidth="75.0" text="Description"/>
+            <TableColumn fx:id="tcAmount" prefWidth="75.0" text="Amount"/>
+        </columns>
+        <columnResizePolicy>
+            <TableView fx:constant="CONSTRAINED_RESIZE_POLICY"/>
+        </columnResizePolicy>
+    </TableView>
+    <HBox alignment="CENTER_LEFT" spacing="10.0">
+        <Label text="Total Balance">
+            <font>
+                <Font name="System Bold" size="16.0"/>
+            </font>
+        </Label>
+        <Label fx:id="lblTotalBalance" text="-">
+            <font>
+                <Font name="System Bold" size="16.0"/>
+            </font>
+        </Label>
+    </HBox>
+</VBox>
+
